// convex/plugins/wca/openapiClient/types.gen.ts
export type RegistrationDataV2 = {
  id: number
  registrant_id: number
  user_id: number
  guests?: number
  user: {
    id: number
    name: string
    gender: string
    country_iso2: string
    wca_id: string
  }
  competing: {
    event_ids: Array<string>
    registration_status?: string
    registered_on?: string
    comment?: string
    admin_comment?: string
  }
  payment?: {
    has_paid?: boolean
    payment_status?: string
    paid_amount_iso?: number
    currency_code?: string
    updated_at?: string
  }
}

export type BaseRegistrationConfig = {
  key: string
  isEditable: boolean
  parameters?: {
    [key: string]: unknown
  }
}

export type RequirementsStepConfig = BaseRegistrationConfig

export type WcifAttemptResult = number

export type WcifQualificationAttemptResult = {
  whenDate: string
  type: string
  resultType: "single" | "average"
  level: WcifAttemptResult
}

export type WcifRanking = number

export type WcifQualificationRanking = {
  whenDate: string
  type: string
  resultType: "single" | "average"
  level: WcifRanking
}

export type WcifQualificationAnyResult = {
  whenDate: string
  type: string
  resultType: "single" | "average"
}

export type WcifQualification =
  | ({
      type: "attemptResult"
    } & WcifQualificationAttemptResult)
  | ({
      type: "ranking"
    } & WcifQualificationRanking)
  | ({
      type: "anyResult"
    } & WcifQualificationAnyResult)

export type WcifPersonalBest = {
  eventId: string
  best: WcifAttemptResult
  worldRanking: number
  continentalRanking: number
  nationalRanking: number
  type: "single" | "average"
}
// ...
export type WcifTimeLimit = {
  centiseconds: number
  cumulativeRoundIds: Array<string>
}

export type WcifCutoff = {
  numberOfAttempts: number
  attemptResult: WcifAttemptResult
}

export type WcifAdvancementConditionRanking = {
  type: string
  level: WcifRanking
}

export type WcifPercent = number

export type WcifAdvancementConditionPercent = {
  type: string
  level: WcifPercent
}

export type WcifAdvancementConditionAttemptResult = {
  type: string
  level: WcifAttemptResult
}

export type WcifAdvancementCondition =
  | ({
      type: "ranking"
    } & WcifAdvancementConditionRanking)
  | ({
      type: "percent"
    } & WcifAdvancementConditionPercent)
  | ({
      type: "attemptResult"
    } & WcifAdvancementConditionAttemptResult)

export type WcifAttempt = {
  result: WcifAttemptResult
  reconstruction?: string
}

export type WcifResult = {
  personId: number
  ranking?: number
  attempts: Array<WcifAttempt>
  best: WcifAttemptResult
  average: WcifAttemptResult
}

export type WcifScramble = string

export type WcifScrambleSet = {
  id: number
  scrambles: Array<WcifScramble>
  extraScrambles: Array<WcifScramble>
}

export type WcifRound = {
  id: string
  format: "1" | "2" | "3" | "5" | "a" | "m"
  timeLimit?: WcifTimeLimit
  cutoff?: WcifCutoff
  advancementCondition?: WcifAdvancementCondition
  results: Array<WcifResult>
  scrambleSetCount: number
  scrambleSets: Array<WcifScrambleSet>
  extensions: Array<unknown>
}
// ...
export type UserAvatar = {
  url: string
  thumb_url?: string
  is_default: boolean
}

export type WcifRegistration = {
  wcaRegistrationId: number
  eventIds: Array<string>
  status: string
  isCompeting: boolean
}

export type WcifAssignment = {
  activityId: number
  stationNumber: number
  assignmentCode: string
}

export type WcifPerson = {
  registrantId: number
  name: string
  wcaUserId: number
  countryIso2: string
  gender: string
  birthdate?: string
  email?: string
  avatar?: UserAvatar
  roles: Array<"delegate" | "trainee-delegate" | "organizer">
  registration?: WcifRegistration
  assignments: Array<WcifAssignment>
  personalBests: Array<WcifPersonalBest>
  extensions: Array<unknown>
}
// ...
export type WcifCountryCode = string

export type WcifActivityCode = string

export type WcifActivity = {
  id: number
  name: string
  activityCode: WcifActivityCode
  startTime: string
  endTime: string
  childActivities: Array<WcifActivity>
  scrambleSetId?: number
  extensions: Array<unknown>
}

export type WcifRoom = {
  id: number
  name: string
  color: string
  activities: Array<WcifActivity>
  extensions: Array<unknown>
}

export type WcifVenue = {
  id: number
  name: string
  latitudeMicrodegrees: number
  longitudeMicrodegrees: number
  countryIso2: WcifCountryCode
  timezone: string
  rooms: Array<WcifRoom>
  extensions: Array<unknown>
}

export type WcifSchedule = {
  startDate: string
  numberOfDays: number
  venues: Array<WcifVenue>
}
// ...
export type WcifEvent = {
  id: string
  rounds: Array<WcifRound>
  competitorLimit?: number
  qualification?: WcifQualification
  extensions: Array<unknown>
}